Added the authordetails to all quote pages, cleaned up some code
