+++ src/components/AuthorDetails/AuthorDetails.jsx
@@ -1,6 +1,4 @@
-import { useDispatch, useSelector } from 'react-redux';
-import { useState } from 'react';
-import { Card, IconButton, CardContent, Typography, Modal } from '@mui/material';
+import { useSelector } from 'react-redux';
 
 
 
@@ -9,24 +7,18 @@ function AuthorDetails() {
 
     console.log('Author Details check', authorDetails);
 
-    const style = {
-        position: 'absolute',
-        top: '50%',
-        left: '50%',
-        transform: 'translate(-50%, -50%)',
-        width: 400,
-        bgcolor: 'background.paper',
-        border: '2px solid #000',
-        boxShadow: 24,
-        p: 4,
-    };
 
     return (
         <div className="authorDetailsDiv">
             {
                 authorDetails.results.map(author => {
                     return (
-                        <p>{author.bio}</p>
+                        <div className='modalDiv'>
+                            <h1>{author.name}</h1>
+                            <p>{author.bio}</p>
+                        </div>
+
+
                     )
                 })
             }
