+++ src/components/Quotes/Quotes.jsx
@@ -2,13 +2,15 @@ import { useDispatch, useSelector } from 'react-redux';
 import { useState } from 'react';
 import { Container, Grid, Card, IconButton, CardContent, Typography, Button, Modal } from '@mui/material';
 import FavoriteIcon from '@mui/icons-material/Favorite';
+import AuthorDetails from '../AuthorDetails/AuthorDetails';
 
 function Quotes(search) {
-    
+
 
     const searchQuotes = useSelector(store => store.searchQuotes);
     const favorites = useSelector((store) => store.quotes.favorites);
-    
+    const [open, setOpen] = useState(false);
+
     let page = 2;
     const dispatch = useDispatch();
 
@@ -28,12 +30,34 @@ function Quotes(search) {
     const nextPage = () => {
         let newPage = page++;
         console.log('this is the new page', newPage);
-        dispatch({ type:'NEXT_PAGE', payload: {search: search, page: newPage}})
+        dispatch({ type: 'NEXT_PAGE', payload: { search: search, page: newPage } })
     }
 
     const showAuthorDetails = (author) => {
         dispatch({ type: 'SET_AUTHOR', payload: author });
+        handleOpen();
     }
+
+    const handleClose = () => setOpen(false);
+    const handleOpen = () => setOpen(true);
+
+    //?Sams alert
+    // const removeFromFavorites = (quote) => {
+    //   alert('Are you sure you want to remove this quote from Favorites?');
+    //   dispatch({ type: 'REMOVE_FROM_FAVORITES', payload: quote });
+    // };
+
+    const style = {
+        position: 'absolute',
+        top: '50%',
+        left: '50%',
+        transform: 'translate(-50%, -50%)',
+        width: 400,
+        bgcolor: 'background.paper',
+        border: '2px solid #000',
+        boxShadow: 24,
+        p: 4,
+    };
 
     return (
         <center>
@@ -72,6 +96,15 @@ function Quotes(search) {
                                                     </Typography>
                                                     <br />
                                                     <Typography variant="h8" id="authorDetail" onClick={() => showAuthorDetails(quote.author)}>{quote.author}</Typography>
+                                                    <Modal open={open}
+                                                        onClose={handleClose}
+                                                    >
+                                                        <Card sx={style}>
+                                                            <center>
+                                                                <AuthorDetails />
+                                                            </center>
+                                                        </Card>
+                                                    </Modal>
                                                     <IconButton
                                                         onClick={() => {
                                                             if (!isInFavorites(quote)) {
@@ -86,20 +119,19 @@ function Quotes(search) {
                                                         />
                                                     </IconButton>
                                                 </CardContent>
-                                               
+
                                             </Card>
                                             <br />
-                                        </Grid> 
+                                        </Grid>
                                     </center>
                                 </Container>
-                                
                             </div>
                         )
                     )
                 }
             </div>
-            <br />
             <Button onClick={nextPage}>Next Page</Button>
+            <br />
         </center>
     )
 } //! End Quotes ()
