+++ src/components/Quotes/SearchQuotes.jsx
@@ -9,12 +9,10 @@ function SearchQuotes() {
     const [search, setSearch] = useState('');
     const dispatch = useDispatch();
 
-
     const handleSubmit = () => {
         dispatch({ type: 'SET_SEARCH', payload: search })
         dispatch({ type: 'SET_AUTHOR', payload: search }) // This is a dispatch to search by authors
     }
-
 
     return (
         <center>
